+++ FBSDKShareKit/FBSDKShareKit/Content/ShareLinkContent.swift
@@ -81,7 +81,7 @@ extension ShareLinkContent: SharingContent {
   }
 }
 
-extension ShareLinkContent: SharingValidation {
+extension ShareLinkContent: SharingValidatable {
   /// Asks the receiver to validate that its content or media values are valid.
   @objc(validateWithOptions:error:)
   public func validate(options bridgeOptions: ShareBridgeOptions) throws {