+++ FBSDKShareKit/FBSDKShareKit/Content/ShareMediaContent.swift
@@ -68,7 +68,7 @@ extension ShareMediaContent: SharingContent {
   }
 }
 
-extension ShareMediaContent: SharingValidation {
+extension ShareMediaContent: SharingValidatable {
   /// Asks the receiver to validate that its content or media values are valid.
   @objc(validateWithOptions:error:)
   public func validate(options bridgeOptions: ShareBridgeOptions) throws {