+++ FBSDKShareKit/FBSDKShareKit/Content/ShareVideo.swift
@@ -104,7 +104,7 @@ extension ShareVideo: DependentAsType {
 
 // MARK: - Validation
 
-extension ShareVideo: SharingValidation {
+extension ShareVideo: SharingValidatable {
   /// Asks the receiver to validate that its content or media values are valid.
   @objc(validateWithOptions:error:)
   public func validate(options bridgeOptions: ShareBridgeOptions) throws {